Refactored auth to authentication to fix var shadowing linter error
