+++ src/js/reducers/authentication.js
